requirements.txt
streamlit
requests
python-dotenv
gTTS
flask
flask-cors
